ParanoidGlyph: Introduce MusicVisualizerService
> Yo girl, come take a look at deez lights

The service uses the AudioFX's Visualizer API to capture the audio data,
which is then processed in the processAudioFFT() method to calculate the
average energy in seperate frequency bands in terms of the spectrum.

The processAudioFFT() method takes in a byte array of FFT data and
calculates the average energy of three frequency bands: low, mid, and
high. The frequency ranges for each band are defined as constants in the
class. The method first calculates the total energy in the low frequency
band by summing the absolute values of the FFT data for the appropriate
indices. It then repeats this process for the other frequency bands.

The method also calculates the average energy of each band over a one
second interval, by keeping track of the total energy and number of
samples in each band for each interval, and dividing the total energy by
the number of samples at the end of the interval. The
mCurrentAvgEnergyOneSec array is used to store the current average
energy for each band, while the mRunningSoundAvg array is used to store
the running total of sound energy in each band.

Co-Authored-By: Fabian Leutenegger <fabian.leutenegger@bluewin.ch>
Change-Id: I1ee43f97a424dc63ae552acc11db755486bf2e61